Externalize mock objects to the __mocks__ folder
