+++ src/Components/ResultsList/ResultsList.test.jsx
@@ -4,46 +4,15 @@ import React from 'react';
 import TestUtils from 'react-dom/test-utils';
 import { MemoryRouter } from 'react-router-dom';
 import ResultsList from './ResultsList';
+import resultsObject from '../../__mocks__/resultsObject';
 
 describe('ResultsListComponent', () => {
   let results = null;
   let wrapper = null;
 
-  const resultsArray = {
-    count: 2,
-    results: [
-      { id: 6,
-        grade: '05',
-        skill: 'OFFICE MANAGEMENT (9017)',
-        bureau: '150000',
-        organization: 'YAOUNDE CAMEROON (YAOUNDE)',
-        position_number: '00025003',
-        is_overseas: true,
-        create_date: '2006-09-20',
-        update_date: '2017-06-08',
-        post: { id: 162, tour_of_duty: '2YRR', code: 'LT6000000', location: 'MASERU, LESOTHO', cost_of_living_adjustment: 0, differential_rate: 15, danger_pay: 0, rest_relaxation_point: 'London', has_consumable_allowance: false, has_service_needs_differential: false },
-        languages: [
-        { id: 1, language: 'French (FR)', written_proficiency: '2', spoken_proficiency: '2', representation: 'French (FR) 2/2' },
-        ],
-      },
-      { id: 60,
-        grade: '03',
-        skill: 'OFFICE MANAGEMENT (9017)',
-        bureau: '150000',
-        organization: 'YAOUNDE CAMEROON (YAOUNDE)',
-        position_number: '00025003',
-        is_overseas: true,
-        create_date: '2006-09-20',
-        update_date: '2017-06-08',
-        post: null,
-        languages: [],
-      },
-    ],
-  };
-
   beforeEach(() => {
     results = TestUtils.renderIntoDocument(<MemoryRouter>
-      <ResultsList results={resultsArray} />
+      <ResultsList results={resultsObject} />
     </MemoryRouter>);
   });
 
@@ -52,12 +21,12 @@ describe('ResultsListComponent', () => {
   });
 
   it('can receive props', () => {
-    wrapper = shallow(<ResultsList results={resultsArray} />);
+    wrapper = shallow(<ResultsList results={resultsObject} />);
     expect(wrapper.instance().props.results.results[0].id).toBe(6);
   });
 
   it('can call the onChildToggle function', () => {
-    wrapper = shallow(<ResultsList results={resultsArray} />);
+    wrapper = shallow(<ResultsList results={resultsObject} />);
     // define the instance
     const instance = wrapper.instance();
     // spy the logout function
