+++ src/Components/ResultsPage/ResultsPage.test.jsx
@@ -3,48 +3,17 @@ import React from 'react';
 import sinon from 'sinon';
 import ResultsPage from './ResultsPage';
 import { POSITION_SEARCH_SORTS, POSITION_PAGE_SIZES } from '../../Constants/Sort';
+import resultsObject from '../../__mocks__/resultsObject';
 
 describe('ResultsPageComponent', () => {
   let wrapper = null;
-
-  const resultsArray = {
-    count: 2,
-    results: [
-      { id: 6,
-        grade: '05',
-        skill: 'OFFICE MANAGEMENT (9017)',
-        bureau: '150000',
-        organization: 'YAOUNDE CAMEROON (YAOUNDE)',
-        position_number: '00025003',
-        is_overseas: true,
-        create_date: '2006-09-20',
-        update_date: '2017-06-08',
-        post: { id: 162, tour_of_duty: '2YRR', code: 'LT6000000', location: 'MASERU, LESOTHO', cost_of_living_adjustment: 0, differential_rate: 15, danger_pay: 0, rest_relaxation_point: 'London', has_consumable_allowance: false, has_service_needs_differential: false },
-        languages: [
-        { id: 1, language: 'French (FR)', written_proficiency: '2', spoken_proficiency: '2', representation: 'French (FR) 2/2' },
-        ],
-      },
-      { id: 60,
-        grade: '03',
-        skill: 'OFFICE MANAGEMENT (9017)',
-        bureau: '150000',
-        organization: 'YAOUNDE CAMEROON (YAOUNDE)',
-        position_number: '00025003',
-        is_overseas: true,
-        create_date: '2006-09-20',
-        update_date: '2017-06-08',
-        post: null,
-        languages: [],
-      },
-    ],
-  };
 
   const defaultSort = '';
   const defaultPageSize = 10;
 
   it('is defined', () => {
     wrapper = shallow(<ResultsPage
-      results={resultsArray}
+      results={resultsObject}
       hasErrored
       isLoading={false}
       sortBy={POSITION_SEARCH_SORTS}
@@ -58,7 +27,7 @@ describe('ResultsPageComponent', () => {
 
   it('can receive props', () => {
     wrapper = shallow(<ResultsPage
-      results={resultsArray}
+      results={resultsObject}
       hasErrored
       isLoading={false}
       sortBy={POSITION_SEARCH_SORTS}
@@ -85,7 +54,7 @@ describe('ResultsPageComponent', () => {
 
   it('can call the onChildToggle function', () => {
     wrapper = shallow(<ResultsPage
-      results={resultsArray}
+      results={resultsObject}
       sortBy={POSITION_SEARCH_SORTS}
       defaultSort={defaultSort}
       pageSizes={POSITION_PAGE_SIZES}
@@ -98,7 +67,7 @@ describe('ResultsPageComponent', () => {
 
   it('can call the queryParamUpdate function', () => {
     wrapper = shallow(<ResultsPage
-      results={resultsArray}
+      results={resultsObject}
       sortBy={POSITION_SEARCH_SORTS}
       defaultSort={defaultSort}
       pageSizes={POSITION_PAGE_SIZES}
